+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.0...HEAD)
+# [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.1...HEAD)
+
+# [v20.0.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.0...v20.0.1)
+### Changed
+* [STREAM-1362](https://inindca.atlassian.net/browse/STREAM-1362) - Retry notification subscription
+* [STREAM-1711](https://inindca.atlassian.net/browse/STREAM-1711) - Fetch current alerting leader status sooner and add retry; upgrade `axios` to v1.17.0 and `ws` to v8.21.0.
 
 # [v20.0.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.8.0...v20.0.0)
 ### Breaking Changes
@@ -24,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [v19.7.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.6.0...v19.7.0)
 ### Added
 * [STREAM-1155](https://inindca.atlassian.net/browse/STREAM-1155) - Allow clients to become the alerting leader and listen for alerting leader events.
+* [STREAM-1003](https://inindca.atlassian.net/browse/STREAM-1003) - Update and modernize documentation.
 
 # [v19.6.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.5.0...v19.6.0)
 ### Added